third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/multiple-enforce-policies-expected.txt

CONSOLE ERROR: line 12: Refused to load the image 'ftp://blah.test/' because it violates the following Content Security Policy directive: "img-src http://* https://*".

CONSOLE ERROR: line 12: Refused to load the image 'ftp://blah.test/' because it violates the following Content Security Policy directive: "img-src http://*".
